fix 修复课程学分类型为Double

+++ bluett-ssms/src/main/java/org/bluett/ssms/domain/Course.java
@@ -37,7 +37,7 @@ public class Course extends BaseEntity {
     /**
      * 学分
      */
-    private Long credit;
+    private Double credit;
     /**
      * 开始日期
      */
